requirements.txt
boto==2.49.0
boto3==1.9.142
botocore==1.12.142
certifi==2019.3.9
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2
djangorestframework==3.9.4
docutils==0.14
gensim==3.7.2
gunicorn==19.9.0
idna==2.8
jmespath==0.9.4
numpy==1.16.3
psycopg2==2.8.2
python-dateutil==2.8.0
pytz==2019.1
requests==2.21.0
s3transfer==0.2.0
scipy==1.2.1
six==1.12.0
smart-open==1.8.3
sqlparse==0.3.0
urllib3==1.24.3
whitenoise==3.3.1
